完成出入库功能；

+++ src/main/java/com/zhannicholas/cpwms/service/impl/RecordInServiceImpl.java
@@ -5,7 +5,8 @@ import com.zhannicholas.cpwms.domain.model.RecordIn;
 import com.zhannicholas.cpwms.domain.model.Respository;
 import com.zhannicholas.cpwms.domain.model.Supplier;
 import com.zhannicholas.cpwms.domain.repository.RecordInRepository;
-import com.zhannicholas.cpwms.service.RecordInService;
+import com.zhannicholas.cpwms.service.*;
+import com.zhannicholas.cpwms.util.DateUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -18,10 +19,18 @@ import java.util.List;
 @Transactional
 public class RecordInServiceImpl implements RecordInService {
     private final RecordInRepository recordInRepository;
+    private final SupplierService supplierService;
+    private final PartsService partsService;
+    private final RespositoryService respositoryService;
+    private final StorageService storageService;
 
     @Autowired
-    public RecordInServiceImpl(RecordInRepository recordInRepository) {
+    public RecordInServiceImpl(RecordInRepository recordInRepository, SupplierService supplierService, PartsService partsService, RespositoryService respositoryService, StorageService storageService) {
         this.recordInRepository = recordInRepository;
+        this.supplierService = supplierService;
+        this.partsService = partsService;
+        this.respositoryService = respositoryService;
+        this.storageService = storageService;
     }
 
     @Override
@@ -46,7 +55,7 @@ public class RecordInServiceImpl implements RecordInService {
 
     @Override
     public RecordIn findOne(int recordId) {
-        return recordInRepository.findByRecordId(recordId);
+        return recordInRepository.findById(recordId);
     }
 
     @Override
@@ -55,7 +64,32 @@ public class RecordInServiceImpl implements RecordInService {
     }
 
     @Override
-    public void delete(int recordId) {
-        recordInRepository.deleteById(recordId);
+    public boolean delete(int recordId) {
+        if(isValidRecordInId(recordId)) {
+            recordInRepository.deleteById(recordId);
+            return true;
+        }
+        return false;
+    }
+
+    @Override
+    public boolean saveRecord(int supplierId, int partsId, int number, String person, int repoId) {
+        if(supplierService.isValidSupplierId(supplierId) &&
+                partsService.isValidPartsId(partsId) &&
+                respositoryService.isValidRepoId(repoId) &&
+                person != null &&
+                number > 0){
+            // 保存到入库表
+            recordInRepository.saveRecord(supplierId, partsId, number, DateUtil.fromUtilDate(), person, repoId);
+            // 更新库存
+            storageService.increaseStorage(partsId, repoId, number);
+            return true;
+        }
+        return false;
+    }
+
+    @Override
+    public boolean isValidRecordInId(int recordInId) {
+        return recordInRepository.findById(recordInId) != null;
     }
 }
